storage: block overwriting unexpected objects
This wasn't documented in Google's JSON API but I found in the XML API
that a `if generation 0` condition means `if doesn't exist`. Now we can
guard against overwriting objects we don't know about similar to how we
already guard against overwriting objects that have unexpectedly
changed. This does impose the new requirement that callers must always
use Fetch before Upload and Copy but that is already done.